refactor: extract RocketChatRoomClient from RocketChatService
Move room operations (getRoomsOfUser, getChatInfo, getChatUsers,
setRoomReadOnly, setRoomWriteable, saveRoomSettings, setRoomState)
to new RocketChatRoomClient. Delegate from RocketChatService.

Remove MongoClient dependency from RocketChatService (now in room
client). RocketChatRoomClient handles MongoDB subscription queries
for getChatUsers.
